+++ CHANGELOG.md
@@ -48,6 +48,7 @@ _The file [UPGRADING.md](UPGRADING.md) contains important information about upgr
 - core: fix buffer selected when searching by partial name and another buffer has a similar name ([#700](https://github.com/weechat/weechat/issues/700))
 - irc: fix memory leak in SASL authentication with mechanism "ecdsa-nist256p-challenge"
 - javascript, lua, perl, php, ruby, tcl: fix return value of function hdata_longlong
+- relay/irc: remove message tags not negotiated by the client in redirected messages ([#2359](https://github.com/weechat/weechat/issues/2359))
 - relay/api: fix notify tag in line when already read on remote ([#2343](https://github.com/weechat/weechat/issues/2343))
 
 ### Security